project_mini/requirements.txt
Django>=4.0,<5.0
djangorestframework
celery
redis
drf-yasg
gunicorn
psycopg2-binary
django-environ
reportlab
djangorestframework-simplejwt
fpdf